Move listening logic to StreamStorage class.
